Move Ouroboros.Storage.ChainDB.Mock/Model to the testsuite
The existing modules with the same names in the testsuite, which contain tests
for the mock and model, are given a `.Test` suffix.
