+++ ouroboros-consensus/test-storage/Test/Ouroboros/Storage/ChainDB.hs
@@ -7,8 +7,8 @@ import           Test.Tasty
 import qualified Test.Ouroboros.Storage.ChainDB.AddBlock as AddBlock
 import qualified Test.Ouroboros.Storage.ChainDB.ImmDB as ImmDB
 import qualified Test.Ouroboros.Storage.ChainDB.Iterator as Iterator
-import qualified Test.Ouroboros.Storage.ChainDB.Mock as Mock
-import qualified Test.Ouroboros.Storage.ChainDB.Model as Model
+import qualified Test.Ouroboros.Storage.ChainDB.Mock.Test as Mock
+import qualified Test.Ouroboros.Storage.ChainDB.Model.Test as Model
 import qualified Test.Ouroboros.Storage.ChainDB.StateMachine as StateMachine
 
 tests :: TestTree
